+++ R/utilities.R
@@ -17,3 +17,23 @@ not_empty <- function(xs) {
     }
   })
 }
+
+#' Return a unit version of the argument.
+#'
+#' @param x Number or unit object.
+#' @return unit(x, "lines") if number or the unchanged argument if it's already
+#'  a unit object.
+#' @noRd
+to_unit <- function(x) {
+  # don't change arg if already unit
+  if (class(x) == "unit") {
+    return(x)
+  }
+
+  # NA used to exclude points from repulsion calculations
+  if (is.na(x)) {
+    return(NA)
+  }
+
+  unit(x, "lines")
+}