+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(ggrepel)
+
+test_check("ggrepel")